always download a new file
If you had a file already downloaded (most likely at
/tmp/chromedriver/chromedriver_linux64.zip) node-chromedriver would not
update the file, but use it, so you were not actually updating the
chromedriver files, only the npm package. This fixed that problem. This
is a quickfix. A better approach would be to add the version number to
the download, like /tmp/chromedriver/chromedriver_linux64_2_10.zip so we
could use the cached file and update when needed.